[Ubuntu] Check netadmin privilege availability before testing podman network

+++ images/ubuntu/scripts/tests/Tools.Tests.ps1
@@ -337,9 +337,13 @@ Describe "Containers" {
 
     # https://github.com/actions/runner-images/issues/7753
     It "podman networking" -TestCases "podman CNI plugins" {
-        "podman network create -d bridge test-net" | Should -ReturnZeroExitCode
-        "podman network ls" | Should -Not -OutputTextMatchingRegex "Error"
-        "podman network rm test-net" | Should -ReturnZeroExitCode
+        if (ip link add dummy0 type dummy) {
+            "ip link delete dummy0" | Should -ReturnZeroExitCode
+
+            "podman network create -d bridge test-net" | Should -ReturnZeroExitCode
+            "podman network ls" | Should -Not -OutputTextMatchingRegex "Error"
+            "podman network rm test-net" | Should -ReturnZeroExitCode
+        }
     }
 
 }
